Use fake read ids to ensure uniqueness
Previously we passed blastn the read id as it was in the input file
However, read ids aren't entirely standardized and could be non-unique
particularly if we strip out everything after the first space. To completely
avoid this problem, we now just pass data to blastn with the read ids replaced
with an incrementing value (1,2,3...).